refactor storage layer
