+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/model/StringSerializer.java
@@ -1,4 +1,4 @@
-package dev.openfeature.contrib.providers.flagd.resolver.process.storage;
+package dev.openfeature.contrib.providers.flagd.resolver.process.model;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.TreeNode;
@@ -7,15 +7,15 @@ import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 
 import java.io.IOException;
 
-public class StringParser extends StdDeserializer<String> {
+class StringSerializer extends StdDeserializer<String> {
 
-    public StringParser() {
+    public StringSerializer() {
         super(String.class);
     }
 
-    @Override public String deserialize(JsonParser p, DeserializationContext ctxt)
+    @Override
+    public String deserialize(JsonParser p, DeserializationContext ctxt)
             throws IOException {
-
         TreeNode node = p.readValueAsTree();
         return node.toString();
     }
