+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/storage/connector/Connector.java
@@ -1,4 +1,4 @@
-package dev.openfeature.contrib.providers.flagd.resolver.process.storage;
+package dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector;
 
 import java.util.function.Consumer;
